fixed error in refgemm: gescal of C with beta was missing

+++ tests/refblas_vs_myblas_simple_4x4_cm_ugemm.cpp
@@ -37,7 +37,7 @@ main() {
 	#endif
 #endif
 
-    typedef size_t Index;
+    typedef long Index;
     typedef GeMatrix<ElementType, Index> MT;
 
     MT A(MAX_SIZE, MAX_SIZE, RowMajor);
